+++ entity/commit/Commit.h
@@ -25,7 +25,7 @@ public:
         checkSum_ = CalculateCheckSum();
     }
 
-    Commit(std::set<File> files, std::string message, int64_t checkSum) :
+    Commit(std::set<File> files, std::string message, size_t checkSum) :
             files_(std::move(files)), message_(std::move(message)), checkSum_(checkSum) {}
 
     Commit(const Commit &rhs) = default;
@@ -47,20 +47,20 @@ public:
     static Commit FromJson(nlohmann::json json);
 
 public:
-    int64_t CalculateCheckSum() {
-        int64_t checkSum = 0;
-        auto size = static_cast<int64_t>(files_.size());
+    size_t CalculateCheckSum() {
+        size_t checkSum = 0;
+        auto size = files_.size();
         for (const auto &file: files_) {
             checkSum += file.Hash() + (size << 2) + (size >> 5);
         }
-        checkSum += static_cast<int64_t>(std::hash<std::string>{}(message_));
+        checkSum += std::hash<std::string>{}(message_);
         return checkSum;
     }
 
 private:
     std::set<File> files_;
     std::string message_;
-    int64_t checkSum_;
+    size_t checkSum_;
 };
